refactor(controller): 관리자 API와 사용자 API를 분리
