+++ src/main/java/roomescape/controller/ReservationAdminController.java
@@ -1,0 +1,21 @@
+package roomescape.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import roomescape.service.ReservationService;
+
+@RestController
+@RequestMapping("/admin/reservations")
+public class ReservationAdminController {
+    private final ReservationService reservationService;
+
+    public ReservationAdminController(ReservationService reservationService) {
+        this.reservationService = reservationService;
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> delete(@PathVariable Long id) {
+        reservationService.delete(id);
+        return ResponseEntity.noContent().build();
+    }
+}
