+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,10 +1,7 @@
 package roomescape.controller;
 
-import java.net.URI;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -48,9 +45,4 @@ public class ReservationController {
                 .body(ReservationResponse.from(reservation));
     }
 
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Long id) {
-        reservationService.delete(id);
-        return ResponseEntity.noContent().build();
-    }
 }
