+++ src/main/java/roomescape/controller/ReservationTimeAdminController.java
@@ -1,0 +1,34 @@
+package roomescape.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import roomescape.domain.ReservationTime;
+import roomescape.dto.ReservationTimeCreateRequest;
+import roomescape.dto.ReservationTimeResponse;
+import roomescape.service.ReservationTimeService;
+
+import static org.springframework.http.HttpStatus.CREATED;
+
+@RestController
+@RequestMapping("/admin/times")
+public class ReservationTimeAdminController {
+    private final ReservationTimeService reservationTimeService;
+
+    public ReservationTimeAdminController(ReservationTimeService reservationTimeService) {
+        this.reservationTimeService = reservationTimeService;
+    }
+
+    @PostMapping
+    public ResponseEntity<ReservationTimeResponse> create(@RequestBody ReservationTimeCreateRequest request) {
+        ReservationTime reservationTime = reservationTimeService.create(request.startAt());
+
+        return ResponseEntity.status(CREATED)
+                .body(ReservationTimeResponse.from(reservationTime));
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> delete(@PathVariable Long id) {
+        reservationTimeService.delete(id);
+        return ResponseEntity.noContent().build();
+    }
+}
