+++ src/main/java/roomescape/dto/ReservationCreateRequest.java
@@ -6,12 +6,9 @@ import java.time.LocalDate;
 
 public record ReservationCreateRequest(
         String name,
-
         @JsonFormat(pattern = "yyyy-MM-dd")
         LocalDate date,
-
         Long timeId,
-
         Long themeId
 ) {
 }
